feat(xpath,xslt,wasm): evaluation budget, tailored collation, WASI runtime proof
PureXML.XPath.Budget is the configurable counterpart of libxml2's
fixed 10M node-set cap (libxml2Compatible preset included), thrown as
a located QueryError instead of a silent refusal, default unbounded.
xsl:sort honors lang with tailored alphabet orders for Polish and
Russian derived from the conformance golds, case-folded. The WASM
gate executes all 985 tests under wasmtime on an actual WASI runtime,
wired into CI; the first run caught a real wasm32 trap (Int is 32
bits there, so format-number's digit extraction trapped at 2^31
despite the 2^53 guard; internals are now Int64 and the xsl:number
bound is platform-correct). Xalan baseline 127 -> 125. Closes #143,
closes #144.

+++ Sources/XSLT/XSLTFunctions.swift
@@ -147,12 +147,14 @@ extension PureXML.XSLT {
         /// here may trap, whatever the value or the picture.
         private static func digits(_ value: Double, scale: Double, places: Int) -> (String, String) {
             let scaled = (value * scale).rounded()
+            // Int64 explicitly: on wasm32 Int is 32 bits and the conversion
+            // would trap at 2^31 (found by the WASI test run, #144).
             guard scaled.isFinite, scaled < 9_007_199_254_740_992, scale < 9_007_199_254_740_992 else {
                 let integer = PureXML.XPath.Value.format(value.rounded(.towardZero))
                 return (integer, String(repeating: "0", count: places))
             }
-            let total = Int(scaled)
-            let divisor = Int(scale)
+            let total = Int64(scaled)
+            let divisor = Int64(scale)
             var fraction = String(total % divisor)
             while fraction.count < places {
                 fraction = "0" + fraction
